Adding new version of the FIST API (#19388)
* renaming dir to try and fix casesensitivity stuff causing validation to fail

* changing the name back since it made more things fail.

Changing the case of the t in IoT to match what the typespec was
generating just made things worse, so I put it back and am going to be
modifying the typespec in a sec here

* changing namespace name to hopefully fix the issues with validation tools

the dir was originally checked in with IoT as the name, and the typespec
was generating a dir with Iot, so the validation tools would fail
because they're being run on a case sentitive os, so it would show as an
entirely new directory

* ran tsp format

* fixing binh operationId

the generated opId wasn't right, so it wasn't associating the example
files with the endpoint correctly

* running tsp format on the binh file

* converting 2024-01-10 from raw swagger to typespec

* updated ts compiler, fixed error with summaryType, ran tsp format

* fixing some of the swagger linter failures

* added supression for summaryName per ts dev suggestion

* removing relatedFiles from SBOM since it wasn't in the 2024-01-10 api

* updates to the examples to fix oav errors

* removing remnants of preview version, fixed api ver in examples and main

* fixing api version on the operations examples

* running prettier

* removing defunct preview version and unused swagger files.

We don't need the swagger files anymore since we're using typespec now

* fixing prop name to not break the sdk for now.

we'll have to make breaking changes to make the arm reviewers happy, but
for now we're just converting what we have to typespec while maintaining
functionality

* another field that needs to be changed to not break the API

* updating examples for prop name changes

* starting new API version for typespec and breaking changes

* adding new states to the ProvisioningState object to match our Status obj

We're possibly going to pivot away from using the status entirely and
start actually using the provisioningState in the future, so starting to
duplicate nonterminal states from the status to the provisioningState

* updating BinH model documentation

* changing various prop names to be more verbose, added a union

* compiling ts files

* adding the cert usage union

* added crypto key type union, mention nist on the flags for insecure usage

* adding a union for cvss scores and adding to cve model

* changing from having a nested item for effective scrore to 2 values

* updated tsp compiler and rebuilt

* running prettier on old example files

* reverting 2024-01-10 api code

Reverting the old swagger to how it was before we started trying to do
the typespec conversion

* reverting readme for swagger version.

reverting the readmen to reference all the swagger files for the last
version of the API we released

* adding new version of the API to the top level readme

* changing tag to match model name.

We renamed CVE to CommonVulnerabilityExposure in several places per ARM
recommendation/request so there was an issue validating the summary type
since I had forgotten to update the union of types

* forgot to recompile after changing the summary tag

* adding supression for linter BodyTopLevelProp check.

There's a bug filed for the linter that is causing false positives that
is causing PRs to fail to be mergable.  adding a supression until the
bug is fixed in the linter

* Update specification/fist/resource-manager/readme.md

Co-authored-by: Mark Cowlishaw <markcowl@microsoft.com>

* Update specification/fist/resource-manager/readme.md

Co-authored-by: Mark Cowlishaw <markcowl@microsoft.com>

* fixing helptext to display the proper tagname

* fixing supression on previous version of the swagger.

Instead of specifying each file, just remove the from clause so it
applies to everything

* removing default value from the RPaaS resource def since we dont want it

* removing unneeded supression.

Since we renamed the props to not conflict, we can remove the supression
line

* changing *Name to *Type to get around a linter error.

Currently, the linter applies a rule to all props ending in Name to have
a pattern property on them. For summaryName since it is an enum it
wasn't making a pattern property on the generated swagger.  We were told
it would be best just to rename the prop so the linter doesn't try to
apply the rule

---------

Co-authored-by: Mike Lucadamo <mlucadamo@microsoft.com>
Co-authored-by: Mark Cowlishaw <markcowl@microsoft.com>